requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
langchain-openai
langchain-chroma
langchain-community
langchain-text-splitters
chromadb
python-dotenv
python-multipart
docx2txt
# ✅ Required for Docx2txtLoader